requirements.txt
astroid==3.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
Click
cryptography==42.0.5
docusign-esign==4.0.0rc1
docusign-rooms==1.3.0
docusign-monitor==1.2.0
docusign-click==1.4.0
docusign-admin==1.4.1
docusign-webforms==1.0.2rc12
docusign-maestro==2.0.0rc1
Flask==2.3.3
Flask-OAuthlib==0.9.6
Flask-Session==0.8.0
flask-wtf==1.2.1
flake8==7.0.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2>=3.1.3
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==2.1.0
pipenv==2023.12.1
py-oauth2==0.0.10
pycparser==2.22
pylint==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests>=2.31.0
requests-oauthlib==1.1.0
six==1.16.0
urllib3>=2.2.1
virtualenv==20.25.3
virtualenv-clone==0.5.7
Werkzeug==2.3.8
wrapt==1.16.0
llama_index
llama_index-llms-deepinfra